move ThemeProvider into new dir - context-providers
