+++ src/components/shared/Frame.js
@@ -1,4 +1,4 @@
-import { useThemeContext } from '../../ThemeProvider';
+import { useThemeContext } from '../../context-providers/ThemeProvider';
 
 function Frame({ title, children }) {
   const { theme } = useThemeContext();
