+++ src/components/shared/Input.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import ThemeProvider from '../../ThemeProvider';
+import ThemeProvider from '../../context-providers/ThemeProvider';
 import themes from '../../themes';
 import Input from './Input';
