+++ src/components/shared/Label.js
@@ -1,4 +1,4 @@
-import { useThemeContext } from '../../ThemeProvider';
+import { useThemeContext } from '../../context-providers/ThemeProvider';
 
 function Label({ text }) {
   const { theme } = useThemeContext();
