+++ src/components/shared/Page.js
@@ -1,4 +1,4 @@
-import { useThemeContext } from '../../ThemeProvider';
+import { useThemeContext } from '../../context-providers/ThemeProvider';
 
 function Page({ title, content }) {
   const { theme } = useThemeContext();
