+++ src/components/shared/SectionP.js
@@ -1,4 +1,4 @@
-import { useThemeContext } from '../../ThemeProvider';
+import { useThemeContext } from '../../context-providers/ThemeProvider';
 
 function SectionP({ text }) {
   const { theme } = useThemeContext();
